requirements.txt
Django==5.2
python-decouple
Pillow
djangorestframework
markdown
django-filter